requirements.txt
flask==1.1.2
cymysql==0.9.1
flask-sqlalchemy==2.3.2
flask-httpauth==3.3.0
flask-wtf==0.14.2
flask-cors==3.0.6
flask-script==2.0.6
requests==2.20.0
wechatpy==1.8.3
gunicorn==19.9.0
email-validator==1.1.0
psutil==5.7.0
qiniu==7.2.8
pylint==2.4.4
pytest==5.4.1
flasgger
flask-admin==1.5.5